Move. 응답 템플릿 패키지 이동
api -> global/api
